% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-person.R
\name{plot_career}
\alias{plot_career}
\title{单人生涯时间线}
\usage{
plot_career(x, eid, show_exposure = TRUE)
}
\arguments{
\item{x}{\link{ukb_worklife} 或 \link{ukb_career} 的返回值。}

\item{eid}{要画的人。}

\item{show_exposure}{是否在下方画暴露强度条。}
}
\value{
一个 ggplot 对象。
}
\description{
年龄轴 × 职业码色带 + gap 状态 + 暴露标记 + \code{career_end} 竖线。
这是"详细个人工作史"最直接的呈现。
}
\examples{
wl <- ukb_worklife(ukb_synth(n = 40), verbose = FALSE)
plot_career(wl, eid = wl$cohort$eid[4])
}
