% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{as_seqdef}
\alias{as_seqdef}
\title{转成 TraMineR 的 \code{stslist}}
\usage{
as_seqdef(x, ...)
}
\arguments{
\item{x}{\link{ukb_states} 的返回值。}

\item{...}{传给 \code{TraMineR::seqdef()}。}
}
\value{
一个 \code{stslist}。
}
\description{
转成 TraMineR 的 \code{stslist}
}
