% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-sample.R
\name{plot_qc}
\alias{plot_qc}
\title{质控六面板}
\usage{
plot_qc(x, bundle = NULL)
}
\arguments{
\item{x}{\link{ukb_worklife} / \link{ukb_career} 的返回值。}

\item{bundle}{\link{ukb_bundle} 的返回值。}
}
\value{
一个 ggplot 对象。
}
\description{
质控六面板
}
