% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-sample.R
\name{plot_type_mix}
\alias{plot_type_mix}
\title{型构成 vs 参考队列}
\usage{
plot_type_mix(x, bundle = NULL)
}
\arguments{
\item{x}{已经过 \link{ukb_typology} 的对象。}

\item{bundle}{给了就同时画参考队列的构成。}
}
\value{
一个 ggplot 对象。
}
\description{
型构成 vs 参考队列
}
